+++ include/cry/prng.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013-2015, Davide Galassi. All rights reserved.
+ * Copyright (c) 2013-2016, Davide Galassi. All rights reserved.
  *
  * This file is part of CRY software.
  *
@@ -18,39 +18,44 @@
  */
 
 /**
- * @file    cry_rand.h
+ * @file    cry_prng.h
  * @brief   Pseudo Random Number Generator
  */
 
-#ifndef _CRY_RAND_H_
-#define _CRY_RAND_H_
+#ifndef _CRY_PRNG_H_
+#define _CRY_PRNG_H_
 
 #ifdef __cplusplus
 extern "C"{
 #endif
 
+#include <stddef.h>
+
 /**
- * Random number generator initialization.
- * Because very often people forget to perform initialization, other rand
+ * Pseudo random number generator initialization.
+ *
+ * Because very often people forget to perform initialization, other prng
  * functions will automatically check to see if this routine has been called
  * and call it at that point if not.
  *
- * @return      0 on success, -1 on failure.
+ * @param   seed        Seed data.
+ * @param   seed_siz    Seed data size.
+ * @return  0 on success, -1 on failure.
  */
-int cry_rand_init(void);
+int cry_prng_init(const unsigned char *seed, size_t seed_siz);
 
 /**
- * Produce a cryptographically strong random sequence.
+ * Produce a pseudo random sequence.
  *
  * @param buf   Destination buffer
  * @param siz   Destination buffer size.
  * @return      0 on success, -1 on failure.
  */
-int cry_rand(unsigned char *buf, unsigned int siz);
+int cry_prng_rand(unsigned char *buf, size_t siz);
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* _CRY_RAND_H_ */
+#endif /* _CRY_PRNG_H_ */