+++ src/prng/prng_posix.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013-2015, Davide Galassi. All rights reserved.
+ * Copyright (c) 2013-2016, Davide Galassi. All rights reserved.
  *
  * This file is part of CRY software.
  *
@@ -17,26 +17,37 @@
  * License along with CRY; if not, see <http://www.gnu/licenses/>.
  */
 
-#include <windows.h>
-#include <wincrypt.h>
+#include "cry/prng.h"
+#include <fcntl.h>
+#include <unistd.h>
+#include <errno.h>
 
-static HCRYPTPROV hProvider;
+static int urand_fd = -1;
 
-int cry_rand_init(void)
+int cry_prng_init(const unsigned char *seed, size_t seed_siz)
 {
-    if (!CryptAcquireContext(&hProvider, 0, 0, PROV_RSA_FULL,
-                             CRYPT_VERIFYCONTEXT)) {
+    urand_fd = open("/dev/urandom", O_RDONLY);
+    if (urand_fd < 0)
         return -1;
-    }
     return 0;
 }
 
-int cry_rand(unsigned char *buf, unsigned int siz)
+int cry_prng_rand(unsigned char *buf, size_t siz)
 {
-    if (!hProvider)
-        cry_rand_init();
-    if (!CryptGenRandom(hProvider, siz, buf))
+    int n;
+
+    if (urand_fd == -1 && cry_prng_init(NULL, 0) != 0)
         return -1;
+
+    while (siz) {
+        if ((n = read(urand_fd, buf, siz)) < 0) {
+            if (errno == EINTR)
+                continue;
+            return -1;
+        }
+        buf += n;
+        siz -= n;
+    }
     return 0;
 }