+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Given a version number MAJOR.MINOR.PATCH
 ------------
 
 - Diffie-Hellman secret exchange
+- OS agnostic CSPRNG with AES-CTR
 
 [0.0.4] - 2016-05-07
 --------------------